move to fluentbit package
