+++ internal/reconciler/logpipeline/fluentbit/status.go
@@ -1,4 +1,4 @@
-package logpipeline
+package fluentbit
 
 import (
 	"context"
