+++ internal/reconciler/logpipeline/fluentbit/sync.go
@@ -1,4 +1,4 @@
-package logpipeline
+package fluentbit
 
 import (
 	"bytes"
