+++ internal/reconciler/logpipeline/fluentbit/sync_test.go
@@ -1,4 +1,4 @@
-package logpipeline
+package fluentbit
 
 import (
 	"context"
